流水码自增改为redis方式 原先数据库方式存在并发问题

+++ springboot/dev-util/src/main/java/org/dev/util/entity/SysSequenceQueue.java
@@ -39,7 +39,7 @@ public class SysSequenceQueue extends TenantEntity {
     /**
      * 当前数值
      */
-    private Integer currentValue;
+    private Long currentValue;
 
     /**
      * 当前数值
